requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
gradio>=4.0.0
numpy>=1.24.0
tqdm>=4.65.0
protobuf>=4.21.0
sentencepiece>=0.1.99
accelerate>=0.25.0
bitsandbytes>=0.41.0 